requirements.txt
fastapi==0.119.0
uvicorn==0.38.0
starlette==0.48.0
python-multipart==0.0.20
pydantic==2.12.3
numpy==2.2.6
pandas==2.3.3
matplotlib==3.10.7
scikit-learn==1.7.2
scipy==1.16.2
opencv-python==4.12.0.88
Pillow==11.3.0
tensorflow-cpu==2.20.0
torch==2.9.0+cpu
torchvision==0.24.0+cpu
torchaudio==2.9.0+cpu
ultralytics==8.3.217
ultralytics-thop==2.0.17
requests==2.32.5
python-dotenv==1.1.1
rich==14.2.0
networkx==3.5
